:recycle: 참여 버튼 suspense 적용 위치 변경
